+++ CHANGES.md
@@ -4,6 +4,9 @@ Unreleased
 - Disable background operations on MacOS and other Unixes where we rely on
   fork. (#8100, fixes #8083, @rgrinberg)
 
+- Add `dune build --dump-gc-stats FILE` argument to dump Garbage Collection
+  stats to a named file. (#8072, @Alizter)
+
 3.9.0 (2023-06-28)
 ------------------